// packages/datasource/src/generated/types/map.schema.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * The schema for map.yaml
 */
export type ProjectMap = {
	building: string;
	x: number;
	y: number;
	floor: {
		name?: string;
		image: string;
		projects: string[];
	}[];
}[];